Fix: ToolBar를 ViewController로부터 분리
